requirements.txt
selenium
python-dotenv
pyautogui
pygetwindow
pillow
pyperclip
pytesseract
